Improve project directory tree
